services/ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Python AI service for game recommendations"
authors = [
    {name = "Nikita Perevalov"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "grpcio (>=1.74.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "grpcio-tools (>=1.74.0,<2.0.0)",
    "deepseek-ai (>=0.0.1,<0.0.2)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
packages = [
    {include = "app"}
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
